vision-api-test-flask/requirements.txt
appnope==0.1.0
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
decorator==4.2.1
Flask==0.12.2
google-api-core==0.1.4
google-auth==1.3.0
google-cloud-vision==0.29.0
googleapis-common-protos==1.5.3
grpcio==1.8.4
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
MarkupSafe==1.0
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
protobuf==3.5.1
ptyprocess==0.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.2.0
pytz==2017.3
requests==2.18.4
rsa==3.4.2
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.1
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
